move setMetadata next to addToMetadata
