+++ packages/forms/experimental/public_api.ts
@@ -11,13 +11,12 @@
  * @description
  * Entry point for all public APIs of this package.
  */
-export * from './src/controls/interop_ng_control';
-export * from './src/controls/control';
-export * from './src/api/structure';
-export * from './src/api/logic';
-export * from './src/api/types';
-export * from './src/api/control';
-export * from './src/api/metadata';
-export * from './src/api/validators';
 export * from './src/api/async';
-export * from './src/api/data';
+export * from './src/api/control';
+export * from './src/api/logic';
+export * from './src/api/metadata';
+export * from './src/api/structure';
+export * from './src/api/types';
+export * from './src/api/validators';
+export * from './src/controls/control';
+export * from './src/controls/interop_ng_control';
