+++ packages/forms/experimental/src/api/logic.ts
@@ -6,7 +6,7 @@
  * found in the LICENSE file at https://angular.dev/license
  */
 
-import {AggregateMetadataKey} from '../api/metadata';
+import {AggregateMetadataKey, MetadataKey} from '../api/metadata';
 import {FieldPathNode} from '../path_node';
 import {assertPathIsCurrent} from '../schema';
 import type {FieldContext, FieldPath, LogicFn, PathKind, TreeValidator, Validator} from './types';
@@ -131,4 +131,15 @@ export function addToMetadata<TValue, TMetadata, TPathKind extends PathKind = Pa
 
   const pathNode = FieldPathNode.unwrapFieldPath(path);
   pathNode.logic.addMetadataRule(key, logic);
+}
+
+export function setMetadata<TValue, TData, TPathKind extends PathKind = PathKind.Root>(
+  path: FieldPath<TValue, TPathKind>,
+  key: MetadataKey<TData>,
+  factory: (ctx: FieldContext<TValue, TPathKind>) => TData,
+) {
+  assertPathIsCurrent(path);
+
+  const pathNode = FieldPathNode.unwrapFieldPath(path);
+  pathNode.logic.addDataFactory(key, factory as (ctx: FieldContext<unknown>) => unknown);
 }
